[#76]refactor:뮤테이션합침,컴포넌트명수정,불필요한훅정리
